oelint: allow usage as a library
by moving the relevant parts into a new file called core

Closes #524
